+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - added
   - Add `action` to `TargetBookmark`
   - Add benchmarks for fetching
+  - Take ignored urls into account in `bogrep import`
 - changed
   - Update to rust 1.75
   - Fix duplicate cache files for `bogrep fetch --urls`